Organize code diff folder
