+++ app/electron/main/code/diff/helpers.ts
@@ -1,0 +1,17 @@
+import { CodeDiffRequest } from '/common/models/code';
+import { TemplateNode } from '/common/models/element/templateNode';
+
+export function createHashedTemplateToCodeDiff(
+    templateToCodeDiff: Map<TemplateNode, CodeDiffRequest>,
+): Map<string, CodeDiffRequest> {
+    const hashedTemplateToCodeDiff = new Map<string, CodeDiffRequest>();
+    for (const [templateNode, codeDiffRequest] of templateToCodeDiff) {
+        const hashedKey = hashTemplateNode(templateNode);
+        hashedTemplateToCodeDiff.set(hashedKey, codeDiffRequest);
+    }
+    return hashedTemplateToCodeDiff;
+}
+
+export function hashTemplateNode(node: TemplateNode): string {
+    return `${node.path}:${node.startTag.start.line}:${node.startTag.start.column}:${node.startTag.end.line}:${node.startTag.end.column}`;
+}
