+++ app/src/lib/editor/engine/copy/index.ts
@@ -24,7 +24,7 @@ export class CopyManager {
         if (selected.length === 0) {
             return;
         }
-        // TODO: Handle multiple
+        // TODO: Handle multiple copies
         const selectedEl = this.editorEngine.elements.selected[0];
         const webviewId = selectedEl.webviewId;
         const webview = this.editorEngine.webviews.getWebview(webviewId);
